Move add_kernel where it is used
